cosmetic changes

+++ HX711.h
@@ -14,7 +14,7 @@ class HX711
 		byte DOUT;		// Serial Data Output Pin
 		byte GAIN;		// amplification factor
 		long OFFSET;	// used for tare weight
-		float SCALE;	// used to return weight in grams, kg, ounces, whatever
+		double SCALE;	// used to return weight in grams, kg, ounces, whatever
 
 	public:
 		// define clock and data pin, channel, and gain factor
@@ -41,17 +41,17 @@ class HX711
 		long read_average(byte times = 10);
 
 		// returns (read_average() - OFFSET), that is the current value without the tare weight; times = how many readings to do
-		double get_value(byte times = 1);
+		long get_value(byte times = 1);
 
 		// returns get_value() divided by SCALE, that is the raw value divided by a value obtained via calibration
 		// times = how many readings to do
-		float get_units(byte times = 1);
+		double get_units(byte times = 1);
 
 		// set the OFFSET value for tare weight; times = how many times to read the tare value
 		void tare(byte times = 10);
 
 		// set the SCALE value; this value is used to convert the raw data to "human readable" data (measure units)
-		void set_scale(float scale = 1.f);
+		void set_scale(double scale = 1.);
 
 		// set OFFSET, the value that's subtracted from the actual reading (tare weight)
 		void set_offset(long offset = 0);
